Detail Info Layout 레이아웃 구현 (화면 하단 구현 필요)

+++ Android/app/src/main/java/com/doubleslas/fifith/alcohol/dto/DetailData.kt
@@ -9,7 +9,7 @@ data class DetailData(
     @SerializedName("highestPrice") val highestPrice: Int,
     @SerializedName("lowestPrice") val lowestPrice: Int,
     @SerializedName("ml") val ml: Int,
-    @SerializedName("abv") val abv: Double,
+    @SerializedName("abv") val abv: Float,
     @SerializedName("starAvg") val starAvg: Float,
     @SerializedName("name") val name: String,
     @SerializedName("kind") val kind: List<String>,
